requirements.txt
# Core dependencies
google-cloud-firestore>=2.13.1
functions-framework>=3.5.0
numpy>=1.26.0
python-dateutil>=2.8.2

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Testing utilities
faker>=20.1.0
